Made all SmartPtr and ConstSmartPtr constructors which take raw pointers explicit and adjusted some code accordingly. Think about the following situation:
void DoNothing(SmartPtr<MyClass> sp)	{}

int main(){
	MyClass* p = new MyClass;
	DoSomething(p);
	delete p;
	return 0;
}

Without explicit constructors, we would perform a double free on p when calling 'delete p' in 'main'! (The first free is performed when sp is destroyed when DoNothing returns).
With explicit constructors, the compiler will already raise an error when one attempts to call DoNothing with a raw pointer.

Note that you may use 'make_sp(rawPtr)' to explicitly construct a smart pointer from a given raw pointer (rawPtr has to have a type! NULL is not a valid argument).

Also note that the default constructor of a smart pointer already initializes it's internal raw pointer to NULL.