Fixed distribution of top-grid to all processes.
Instead of a broadcast of the top-subset from proc 0 to all other processes,
an allgather is now performed. Every process thus contributes his top-subset
to the global top-subset which, again, is stored on all processes.
Ghost elements are considered this time.

This prevents a bug where the lsf is initialized relative to the top-subset
after a couple of projected refinements. In this case it was likely that
proc 0 only contained a part of the top-subset on the top level.